requirements-modelling.txt
absl-py
dm-haiku
dm-tree
jax
jaxlib
matplotlib
ml_collections
numpy
optax
pandas
requests
scipy
tensorflow
